orchestrator/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
langchain-openai>=0.1.8
langchain-core>=0.2.0
langchain>=0.2.0
botbuilder-core>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
pydantic>=2.0.0
python-dotenv>=1.0.0